Some Modification authentication and authorisation done

+++ controllers/moviController.js
@@ -14,26 +14,33 @@ const OMDB_API_KEY = "8cb4187"; // http://www.omdbapi.com/?i=tt3896198&apikey=8c
 
 const saveMovie = async function (req, res) {
   try {
-    const adminId = req.params.adminId;
+    let userId =  req.userId ;
 
-    const checkAdmin = await userModel.findById(adminId);
-
+    const checkAdmin = await userModel.findById(userId);
+console.log(checkAdmin)
     if (checkAdmin.role !== "admin") {
       return res
         .status(400)
         .send({ status: false, message: "Only Admin can upload movie" });
     }
+      
+    if(checkAdmin.isDeleted == true){
+      return res
+        .status(400)
+        .send({ status: false, message: " Admin is not exist" });
+    }
+
     let dataB = req.body;
 
     let { title, year, imdbID, type, poster, adminID } = dataB; // removeable
 
-    adminID = dataB.adminID = adminId
+    adminID = dataB.adminID = userId
     
     const response = await axios.get(
-      `http://www.omdbapi.com/?i=tt3896198&apikey=${OMDB_API_KEY}`
+      `http://www.omdbapi.com/?t=${req.body.title}&apikey=${OMDB_API_KEY}`
     );
     const data = response.data;
-
+console.log(data)
     // console.log(data);
     const movie = {
       title: data.Title,
@@ -41,7 +48,7 @@ const saveMovie = async function (req, res) {
       imdbID: data.imdbID,
       type: data.Type,
       poster: data.Poster, 
-      adminID:adminId
+      adminID:userId
     };
 console.log(movie)
     const moviData = await movieModel.create(movie);
@@ -52,8 +59,8 @@ console.log(movie)
         message: "Movie saved successfully",
         data: moviData,
       });
-  } catch (error) {
-    res.status(500).json({ error: "Internal server error" });
+  } catch (err) {
+    res.status(500).send({status:false , message:err.message});
   }
 };
 
@@ -61,10 +68,12 @@ console.log(movie)
 
 //  ===================================== get movie by Query if no query showing all movie ===========================================
 
-const getMovie = async function(req,res){
+const getMovie = async function(req,res){  
 
-  const userId = req.body.userId
+  let userId =  req.userId 
+
 const data = req.query
+
 const {title,year,type } = data
 if(!title || !year || !type){
   const movies = await movieModel.find({isDeleted:false})
@@ -85,21 +94,10 @@ return res.status(200).send({status:true , data:Query})
 
 const getMovieById = async function(req, res){
 
-  const userId = req.params.userId
   const movieId = req.params.movieId
-  let data = req.body
-let {userID , movies} = data
-  userID = data.userID = userId
-   movies = data.movies = movieId
-  const getMovie = await movieModel.findById(movieId)
-  if(getMovie.isDeleted==false){return res.status(400).send({status:false , message:"this movie does not exist"})}
- 
-const checkWatchList = await watchListModel.findOne({userID:userId, movies:movieId})
-if(!checkWatchList){
-  // here i am adding movies in watchList for each different user 
-  await watchListModel.create(data)
   
-}
+  const getMovie = await movieModel.findById(movieId)
+ 
 const revieweDetails = await reviewModel.find({movies:movieId})
 
 const details = {
@@ -109,6 +107,7 @@ movie: getMovie,
 review:revieweDetails
 
 }
+
 return res.status(200).send({status:false , data:details})
 
 }
@@ -118,22 +117,35 @@ return res.status(200).send({status:false , data:details})
 //==================================  delete movies ====================================
 
 const deleteMovie = async function(req,res){  
-  // admin can be more Than one so i have to authorise the user who
-  //  saved the video only that admin can  delete the movie
 
-  const userId = req.params.userId
+  let userId =  req.userId 
+console.log(userId)
   const movieId = req.params.movieId
+
+// Checking admin or not
 const checkUser = await userModel.findById(userId)
+
 if(checkUser.role != "admin"){return res.status(400).send({status:false , message:"Only admin can delete movie"})}
 
+if(checkUser.isDeleted == true){return res.status(400).send({status:false , message:" user not found"})}
+
+// checking same user or not 
+
+const authorization = await movieModel.findById(movieId)
+
+if(authorization.adminID != userId){return res.status(403).send({status:false , message:"Unathorized Access" })}
+if(authorization.isDeleted == true){return res.status(400).send({status:false , message:" movie doesn't exist "})}
+// soft delete of movie
 await movieModel.findOneAndUpdate({_id:movieId,isDeleted:false},{isDeleted:true},{new:true})
+
+// deleting from watchList 
  await watchListModel.findOneAndDelete({movies:movieId})
-// agr video hi delete ho gya h toh watchId/ wachlist se bhi delete ho jayegi...
-await reviewModel.findOneAndDelete({movies:movieId}) // yaha find hi nhi hogi store q krna
+// deleting review of that movie from db
+await reviewModel.findOneAndDelete({movies:movieId}) 
 
 res.status(200).send({status:true , message:"deleted successfully"})
 
-}
+}
