+++ controllers/revieweController.js
@@ -7,40 +7,49 @@ const reviewModel = require("../models/revieweModel")
 //=============================review creation =============================
 
 
-const reviewdata = async function(req,res){
-    const userId = req.params.userId
+const reviewdata = async function(req, res){
 
-    const data = req.body
+    let userId = req.userId
 
-const {review, rating, movies} = data
+    let movieId = req.params.movieId
 
-const checkReviews = await reviewModel.findOne({userID:userId , movies:movies})
+    let data = req.body
+
+let {review, rating, userID, movies} = data
+// have to validate rating ...
+const checkReviews = await reviewModel.findOne({userID:userId , movies:movieId})
 
 if(checkReviews){return res.status(400).send({status:false, message:"already reviwed"})}
 
-const checkMovie = await movieModel.findById(movies)
+const checkMovie = await movieModel.findById(movieId)
 
 if(checkMovie.isDeleted == true){return res.status(400).send({status:false, message:"movies does not exist"})}
+
 userID = data.userID = userId
+
+movies = data.movies = movieId
 
 const saveReviwes = await reviewModel.create(data)
 
-res.status(201).send({status:true ,message:"Review created" ,data:saveReviwes})
+res.status(201).send({status:true ,message:"Review created successfully" , data:saveReviwes})
 
 
 }
 //========================== get Particular review details =========================
 
 const getParticular = async function(req,res){
-    const userId = req.params.userId
-const movies = req.body.movies
 
-const getMovie = await movieModel.findById(movies)
+let userId = req.userId
+console.log(userId)
+const movieId = req.params.movieId
+
+const getMovie = await movieModel.findById(movieId)
+console.log(getMovie)
 if(getMovie.isDeleted == true){
     return res.status(400).send({status:false , message:"movie does not exist"})
 }
 
-const reviews = await reviewModel.findOne({userID:userId, movies:movies})
+const reviews = await reviewModel.findOne({userID:userId, movies:movieId})
 
 const ReviewData = {
     movies:getMovie,
@@ -56,7 +65,7 @@ res.status(200).send({status:true, data:ReviewData})
 
 
 const getAllReviews = async function(req,res){
-    const userId = req.params.userId
+    let userId = req.userId
 
 const reviews = await reviewModel.find({userID:userId}).populate("movies")
 
@@ -70,17 +79,25 @@ res.status(200).send({status:true, data:reviews})
 
 const updateReview = async function(req, res){
     
-    const userId = req.params.userId
+    let userId = req.userId
+let movieId = req.params.movieId
 
     let data = req.body
 
-let {review, rating, movies} = data
+let {review, rating} = data
+ 
 
-const checkMovie = await movieModel.findById(movies)
+const checkMovie = await movieModel.findById(movieId)
 
 if(checkMovie.isDeleted == true){return res.status(400).send({status:false, message:"movies does not exist"})}
 
-let updateReviews = await reviewModel.findOneAndUpdate({userID:userId , movies:movies},
+const check = await reviewModel.findOne({movies:movieId, userID:userId})
+console.log(check)
+if(!check){return res.status(404).send({status:false , message:"No review found"})}
+
+
+
+let updateReviews = await reviewModel.findOneAndUpdate({userID:userId , movies:movieId},
     {$set:{review:data.review , rating:data.rating}},{new:true})
 
 res.status(200).send({status:true ,message:"Review updated" , data:updateReviews})
@@ -91,12 +108,15 @@ res.status(200).send({status:true ,message:"Review updated" , data:updateReviews
 // ======================== delete Particular review ====================
 
 const DeleteParticular = async function(req,res){
-    const userId = req.params.userId
+    let userId = req.userId
 
-    const movies = req.body.movies
+    const movieId = req.params.movieId
 
-     await reviewModel.findOneAndDelete({userID:userId , movies:movies},{new:true})
+    let check =    await reviewModel.findOne({userID:userId , movies:movieId})
+    if(!check){return res.status(404).send({status:false , message:"No review found to delete"})}
+   await reviewModel.findOneAndDelete({userID:userId , movies:movieId},{new:true})
 
+  
 
   res.status(200).send({status:true ,message:"Selected Review deleted" })
 
@@ -104,7 +124,10 @@ const DeleteParticular = async function(req,res){
 //======================== delete all reviews =======================
 
 const DeleteAll = async function(req,res){
-    const userId = req.params.userId
+let userId = req.userId
+
+ const check = await reviewModel.findOne({userID:userId},{new:true})
+ if(!check){return res.status(404).send({status:false , message:" No review found to delete"})}
 
    await reviewModel.deleteMany({userID:userId},{new:true})
